# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: my-car API
  version: 1.0.0
  description: REST API for the my-car rewrite.
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
- url: /api/v1
security:
- cookieSession: []
paths:
...
  /cars:
    get:
      tags:
      - Cars
      operationId: listCars
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Cars ordered by creation time then ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ServerError'
    post:
      tags:
      - Cars
      operationId: createCar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarInput'
      responses:
        '201':
          description: Created car.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ServerError'
  /cars/{carId}:
    parameters:
    - $ref: '#/components/parameters/CarId'
    get:
      tags:
      - Cars
      operationId: getCar
      responses:
        '200':
          description: Car.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ServerError'
    patch:
      tags:
      - Cars
      operationId: updateCar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarUpdate'
      responses:
        '200':
          description: Updated car.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
      - Cars
      operationId: deleteCar
      responses:
        '204':
          description: Car and associated expenses deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: my_car_session
      description: HttpOnly, Secure, SameSite session cookie issued only by the backend.
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor from nextCursor.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    CarId:
      name: carId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource absent or not owned by caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    UUID:
      type: string
      format: uuid
...
    NullableDecimal:
      type:
      - string
      - 'null'
      pattern: ^-?[0-9]+(\.[0-9]+)?$
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          example: validation_failed
        message:
          type: string
        fields:
          type: object
          description: Validation details keyed by request member name (`_body`
            for the request as a whole). Values are machine-readable reasons,
            such as `read_only` (the member cannot be changed), `unknown` (the
            member is not part of the request schema), `invalid_type` (wrong
            JSON type), `too_long` (exceeds the documented maximum length),
            `required` (a required member is missing), `empty` (a string is
            blank after trimming), `invalid_enum` (not one of the documented
            codes), `invalid_date` (not an RFC 3339 full-date),
            `invalid_decimal` (not a decimal string), and `negative` (a
            numeric value below zero).
          additionalProperties:
            type: string
...
    CarInput:
      type: object
      additionalProperties: false
      required:
      - type
      - make
      - name
      - fuel
      properties:
        type:
          type: string
        make:
          type: string
        name:
          type: string
        fuel:
          type: string
          enum:
          - other
          - diesel
          - gasoline
          - electric
        firstRegistration:
          type:
          - string
          - 'null'
          format: date
        licensePlate:
          type:
          - string
          - 'null'
        fin:
          type:
          - string
          - 'null'
        isActive:
          type: boolean
          default: true
        purchaseDate:
          type:
          - string
          - 'null'
          format: date
        purchasePrice:
          $ref: '#/components/schemas/NullableDecimal'
    CarUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        type:
          type: string
        make:
          type: string
        name:
          type: string
        fuel:
          type: string
          enum:
          - other
          - diesel
          - gasoline
          - electric
        firstRegistration:
          type:
          - string
          - 'null'
          format: date
        licensePlate:
          type:
          - string
          - 'null'
        fin:
          type:
          - string
          - 'null'
        isActive:
          type: boolean
        purchaseDate:
          type:
          - string
          - 'null'
          format: date
        purchasePrice:
          $ref: '#/components/schemas/NullableDecimal'
    Car:
      type: object
      description: A car. Every member is always present;
        `firstRegistration`, `licensePlate`, `fin`, `purchaseDate`, and
        `purchasePrice` are `null` when not recorded.
      required:
      - id
      - type
      - make
      - name
      - fuel
      - firstRegistration
      - licensePlate
      - fin
      - isActive
      - purchaseDate
      - purchasePrice
      - createdAt
      - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        type:
          type: string
        make:
          type: string
        name:
          type: string
        fuel:
          type: string
          enum:
          - other
          - diesel
          - gasoline
          - electric
        firstRegistration:
          type:
          - string
          - 'null'
          format: date
        licensePlate:
          type:
          - string
          - 'null'
        fin:
          type:
          - string
          - 'null'
        isActive:
          type: boolean
        purchaseDate:
          type:
          - string
          - 'null'
          format: date
        purchasePrice:
          $ref: '#/components/schemas/NullableDecimal'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CarPage:
      type: object
      required:
      - items
      - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Car'
        nextCursor:
          type:
          - string
          - 'null'
...